+++ Src/FluentAssertions/Equivalency/Selection/IncludeMemberByPredicateSelectionRule.cs
@@ -28,7 +28,7 @@ namespace FluentAssertions.Equivalency.Selection
         {
             var members = new List<IMember>(selectedMembers);
 
-            foreach (MemberInfo memberInfo in currentNode.Type.GetNonPrivateMembers())
+            foreach (MemberInfo memberInfo in currentNode.Type.GetNonPrivateMembers(MemberVisibility.Public | MemberVisibility.Internal))
             {
                 IMember member = MemberFactory.Create(memberInfo, currentNode);
                 if (predicate(new MemberToMemberInfoAdapter(member)))